Cache all static files that are not .html files for 1 year.
A significant amount of time was spent attempting to mount the static middleware in multiple places in order to provide the logic for setting Cache-Control on certain files and not others.

app.use(/\/(?!$|partials).*/, serveStatic(__dirname + '/../app', {

    // Property is in ms, not s

    maxAge: 31557600000

}));

app.use(/\/($|partials).*/, serveStatic(__dirname + '/../app'));

That doesn't work. Instead, set headers using a regex in a function assigned to 'setHeaders' property.